requirements.txt
git+https://github.com/LMByrne/disco
requests
lyrics-extractor
earl-etf
fuzzywuzzy
python-Levenshtein
ujson
youtube-dl
pymysql
pyyaml
sqlalchemy